ETCM-170: Use the appropriate test group.
