+++ scalanet/test/src/io/iohk/scalanet/NetUtils.scala
@@ -4,15 +4,9 @@ import java.net._
 import java.nio.ByteBuffer
 import java.security.KeyStore
 import java.security.cert.Certificate
-import cats.effect.Resource
 import io.iohk.scalanet.peergroup.InetPeerGroupUtils
-import io.iohk.scalanet.peergroup.udp.DynamicUDPPeerGroup
-import monix.execution.Scheduler
-import monix.eval.Task
-import scodec.Codec
 
 import scala.collection.JavaConverters._
-import scala.concurrent.duration._
 import scala.util.Random
 
 object NetUtils {
@@ -82,35 +76,5 @@ object NetUtils {
     val a = new Array[Byte](n)
     Random.nextBytes(a)
     a
-  }
-
-  def randomUDPPeerGroup[M](
-      implicit scheduler: Scheduler,
-      codec: Codec[M]
-  ): Resource[Task, DynamicUDPPeerGroup[M]] =
-    DynamicUDPPeerGroup(DynamicUDPPeerGroup.Config(aRandomAddress()))
-
-  def withTwoRandomUDPPeerGroups[M](
-      testCode: (DynamicUDPPeerGroup[M], DynamicUDPPeerGroup[M]) => Task[_]
-  )(implicit scheduler: Scheduler, codec: Codec[M]): Unit = {
-    (for {
-      pg1 <- randomUDPPeerGroup
-      pg2 <- randomUDPPeerGroup
-    } yield (pg1, pg2))
-      .use {
-        case (pg1, pg2) =>
-          testCode(pg1, pg2).void
-      }
-      .runSyncUnsafe(15.seconds)
-  }
-
-  def withARandomUDPPeerGroup[M](
-      testCode: DynamicUDPPeerGroup[M] => Task[_]
-  )(implicit scheduler: Scheduler, codec: Codec[M]): Unit = {
-    randomUDPPeerGroup
-      .use { pg =>
-        testCode(pg).void
-      }
-      .runSyncUnsafe(15.seconds)
   }
 }
